apps/ai/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
langchain==0.1.9
unstructured[all-docs]==0.12.5
sentence-transformers==3.0.0
llama-index-core==0.10.27
llama-index-readers-file==0.1.22
llama-index-llms-langchain==0.1.3
llama-index-embeddings-langchain==0.1.2
llama-index-vector-stores-milvus==0.1.6
llama-index-vector-stores-postgres==0.1.5
pymilvus==2.4.0
dataclass-wizard==0.22.3
opencv-python==4.8.0.74
minio==7.2.5
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
langchain-core==0.1.29
langchain-nvidia-ai-endpoints==0.1.1
opentelemetry-sdk==1.23.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
faiss-cpu==1.7.4
httpcore==1.0.2
httpx==0.26.0
bleach==6.1.0
flatdict==4.0.1
chardet==5.2.0
gradio==4.13.0
jinja2==3.1.3
numpy==1.26.4
protobuf==4.25.3
pycountry==23.12.11
tritonclient[all]==2.43.0